+++ src/index.ts
@@ -24,11 +24,11 @@ import * as data from './tools/data/index.js';
 import * as users from './tools/users/index.js';
 import * as testing from './tools/testing/index.js';
 import * as metadata from './tools/metadata/index.js';
+import * as dynamic from './tools/dynamic/index.js';
 import Cache from './shared/cache.js';
 import { Telemetry } from './telemetry.js';
 import { SfMcpServer } from './sf-mcp-server.js';
-
-const TOOLSETS = ['all', 'testing', 'orgs', 'data', 'users', 'metadata', 'experimental'] as const;
+import { determineToolsetsToEnable, TOOLSETS } from './shared/tools.js';
 
 /**
  * Sanitizes an array of org usernames by replacing specific orgs with a placeholder.
@@ -89,12 +89,12 @@ You can also use special values to control access to orgs:
       },
     }),
     toolsets: Flags.option({
-      options: TOOLSETS,
+      options: ['all', ...TOOLSETS] as const,
       char: 't',
       summary: 'Toolset to enable',
       multiple: true,
       delimiter: ',',
-      default: ['all'],
+      exclusive: ['dynamic-toolsets'],
     })(),
     version: Flags.version(),
     'no-telemetry': Flags.boolean({
@@ -102,6 +102,11 @@ You can also use special values to control access to orgs:
     }),
     debug: Flags.boolean({
       summary: 'Enable debug logging',
+    }),
+    'dynamic-tools': Flags.boolean({
+      summary: 'Enable dynamic toolsets',
+      char: 'd',
+      exclusive: ['toolsets'],
     }),
   };
 
@@ -127,7 +132,7 @@ You can also use special values to control access to orgs:
 
     if (!flags['no-telemetry']) {
       this.telemetry = new Telemetry(this.config, {
-        toolsets: flags.toolsets.join(', '),
+        toolsets: (flags.toolsets ?? []).join(', '),
         orgs: sanitizeOrgInput(flags.orgs),
       });
 
@@ -139,7 +144,7 @@ You can also use special values to control access to orgs:
       });
     }
 
-    Cache.getInstance().set('allowedOrgs', new Set(flags.orgs));
+    await Cache.safeSet('allowedOrgs', new Set(flags.orgs));
     this.logToStderr(`Allowed orgs:\n${flags.orgs.map((org) => `- ${org}`).join('\n')}`);
     const server = new SfMcpServer(
       {
@@ -150,24 +155,39 @@ You can also use special values to control access to orgs:
           tools: {},
         },
       },
-      { telemetry: this.telemetry }
+      {
+        telemetry: this.telemetry,
+        dynamicTools: flags['dynamic-tools'] ?? false,
+      }
     );
 
-    const enabledToolsets = new Set(flags.toolsets);
-    const all = enabledToolsets.has('all');
+    const toolsetsToEnable = determineToolsetsToEnable(flags.toolsets ?? ['all'], flags['dynamic-tools'] ?? false);
 
     // ************************
     // CORE TOOLS (always on)
+    // If you're adding a new tool to the core toolset, you MUST add it to the `CORE_TOOLS` array in shared/tools.ts
+    // otherwise SfMcpServer will not register it.
+    //
+    // Long term, we will want to consider a more elegant solution for registering core tools.
     // ************************
     this.logToStderr('Registering core tools');
     // get username
     core.registerToolGetUsername(server);
     core.registerToolResume(server);
 
+    // DYNAMIC TOOLSETS
+    // ************************
+    if (toolsetsToEnable.dynamic) {
+      this.logToStderr('Registering dynamic tools');
+      // Individual tool management
+      dynamic.registerToolEnableTool(server);
+      dynamic.registerToolListTools(server);
+    }
+
     // ************************
     // ORG TOOLS
     // ************************
-    if (all || enabledToolsets.has('orgs')) {
+    if (toolsetsToEnable.orgs) {
       this.logToStderr('Registering org tools');
       // list all orgs
       orgs.registerToolListAllOrgs(server);
@@ -176,7 +196,7 @@ You can also use special values to control access to orgs:
     // ************************
     // DATA TOOLS
     // ************************
-    if (all || enabledToolsets.has('data')) {
+    if (toolsetsToEnable.data) {
       this.logToStderr('Registering data tools');
       // query org
       data.registerToolQueryOrg(server);
@@ -185,7 +205,7 @@ You can also use special values to control access to orgs:
     // ************************
     // USER TOOLS
     // ************************
-    if (all || enabledToolsets.has('users')) {
+    if (toolsetsToEnable.users) {
       this.logToStderr('Registering user tools');
       // assign permission set
       users.registerToolAssignPermissionSet(server);
@@ -194,16 +214,16 @@ You can also use special values to control access to orgs:
     // ************************
     // testing TOOLS
     // ************************
-    if (all || enabledToolsets.has('testing')) {
+    if (toolsetsToEnable.testing) {
       this.logToStderr('Registering testing tools');
-      testing.registerToolRunApexTest(server);
-      testing.registerToolRunAgentTest(server);
+      testing.registerToolTestApex(server);
+      testing.registerToolTestAgent(server);
     }
 
     // ************************
     // METADATA TOOLS
     // ************************
-    if (all || enabledToolsets.has('metadata')) {
+    if (toolsetsToEnable.metadata) {
       this.logToStderr('Registering metadata tools');
       // deploy metadata
       metadata.registerToolDeployMetadata(server);
@@ -217,8 +237,9 @@ You can also use special values to control access to orgs:
     // This toolset needs to be explicitly enabled ('all' will not include it)
     // Tools don't need to be in an 'experimental' directory, only registered here
     // ************************
-    if (enabledToolsets.has('experimental')) {
+    if (toolsetsToEnable.experimental) {
       this.logToStderr('Registering experimental tools');
+      orgs.registerToolOrgOpen(server);
       // Add any experimental tools here
     }